refactor: Rename to GradeManageSystem and add separate GradeManager class as a wrapper
